Refactored as per the comments
