--- src/_variables.scss
@@ -1,1 +1,0 @@
-$icon-font-path: '../node_modules/bootstrap-sass/assets/fonts/bootstrap/';
