+++ src/app/app.module.ts
@@ -2,11 +2,11 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { BreadcrumbsModule } from './breadcrumbs/breadcrumbs.module';
-import { FooterModule } from './footer/footer.module';
-import { HeaderModule } from './header/header.module';
-import { CoursesPageModule } from './pages/courses-page/courses-page.module';
+import { AppComponent } from './appComponent/app.component';
+import { BreadcrumbsModule } from './breadCrumbsModule/breadcrumbs.module';
+import { FooterModule } from './components/footerComponent/footer.module';
+import { HeaderModule } from './headerModule/header.module';
+import { CoursesPageModule } from './coursesPageModule/courses-page.module';
 
 @NgModule({
   declarations: [
