+++ src/app/appComponent/app.component.scss
