+++ src/app/components/footerComponent/footer.component.html
@@ -1,4 +1,4 @@
-<div class="d-flex justify-content-end">
+<div class="d-flex justify-content-end footer-container">
     <p class="m-3">Copyright Video Courses! All rights reserved!</p>
 </div>
