+++ src/app/components/footerComponent/footer.component.scss
@@ -1,0 +1,5 @@
+app-footer {
+    .footer-container {
+        color: #FFFFFF;
+    }
+}
