+++ src/app/components/footerComponent/footer.component.spec.ts
