+++ src/app/components/footerComponent/footer.component.ts
