+++ src/app/components/logoComponent/logo.component.scss
