+++ src/app/components/logoComponent/logo.component.spec.ts
