+++ src/app/components/logoComponent/logo.component.ts
