+++ src/app/components/logoComponent/logo.module.ts
