+++ src/app/coursesPageModule/components/courseItemComponent/course-item.component.scss
