+++ src/app/coursesPageModule/courses-page.component.scss
