+++ src/app/coursesPageModule/courses-page.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CoursesPageComponent } from './courses-page.component';
-import { CourseItemComponent } from './course-item/course-item.component';
+import { CourseItemComponent } from './components/courseItemComponent/course-item.component';
 import { FormsModule } from '@angular/forms';
