+++ src/app/coursesPageModule/interfaces/course.model.ts
