+++ src/app/headerModule/header.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './header.component';
-import { LogoModule } from '../logo/logo.module';
+import { LogoModule } from '../components/logoComponent/logo.module';
 
 @NgModule({
   declarations: [HeaderComponent],
