+++ src/styles.scss
@@ -1,3 +1,3 @@
 /* You can add global styles to this file, and also import other style files */
-@import 'variables';
+$icon-font-path: '../node_modules/bootstrap-sass/assets/fonts/bootstrap/';
 @import '../node_modules/bootstrap/scss/bootstrap';
